pyproject.toml
[project]
name = "nautilus"
version = "0.1.0"
description = "Nautilus data broker + reasoning engine: intent-aware scoped query execution, multi-agent handoffs, LLM intent analysis, and forensic escalation via Fathom rules"
readme = "README.md"
requires-python = ">=3.14"
license = "MIT"
# TODO: add author email once the maintainer publishes one.
authors = [{ name = "se-jo-ma" }]
keywords = ["fathom", "policy", "broker", "security", "scope", "clips"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "fathom-rules>=0.3.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "pyyaml>=6.0",
    "pyjwt>=2.12.1",
    "cryptography>=46.0.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "elasticsearch>=8",
    "neo4j>=5",
    "mcp>=1.12",
    "sse-starlette>=2.0",
    "python-multipart>=0.0.9",
    "influxdb-client>=1.40",
    "aiobotocore>=2.12",
]

[project.urls]
# TODO: update once the public repository URL is finalized.
Repository = "https://github.com/se-jo-ma/nautilus"
Documentation = "https://github.com/se-jo-ma/nautilus/blob/main/specs/core-broker/design.md"
Issues = "https://github.com/se-jo-ma/nautilus/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "testcontainers[postgres]>=4.0",
    "ruff>=0.5",
    "pyright>=1.1.370",
    "pip-licenses>=5.0",
    "respx>=0.21",
    "pytest-recording>=0.13",
    "testcontainers[elasticsearch,neo4j]>=4.0",
    "locust>=2.0",
    "mkdocs>=1.6",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
    "copier>=9.0",
]
otel = [
    "opentelemetry-sdk>=1.20",
    "opentelemetry-api>=1.20",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
    "prometheus-client>=0.20",
]
llm-anthropic = ["anthropic>=0.40"]
llm-openai = ["openai>=1.50"]

[project.entry-points."fathom.packs"]
data-routing-nist = "nautilus.rule_packs.data_routing_nist"
data-routing-hipaa = "nautilus.rule_packs.data_routing_hipaa"

[project.entry-points."nautilus.adapters"]
influxdb = "nautilus.adapters.influxdb"
s3 = "nautilus.adapters.s3"

[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["nautilus*"]
exclude = ["tests*", "specs*", "docs*"]

[tool.setuptools.package-data]
"nautilus.analysis.llm.prompts" = ["*.txt"]
"nautilus.rules" = ["**/*.yaml"]
"nautilus.ui" = ["templates/**/*", "static/*"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM", "ASYNC"]

[tool.pyright]
strict = ["nautilus", "tests"]
include = ["nautilus", "tests"]
extraPaths = ["packages/nautilus-adapter-sdk/src"]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "slow: slow tests (e.g. latency harness with 1000 requests)",
    "docker: tests that require a local Docker daemon",
]
addopts = "--strict-markers --cov=nautilus --cov-branch"
